+++ package.json
@@ -18,6 +18,10 @@
     ],
     "snapshotSerializers": [
       "<rootDir>/node_modules/enzyme-to-json/serializer"
+    ],
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx}",
+      "!src/index.{js,jsx}"
     ]
   },
   "devDependencies": {
