+++ src/index.js
@@ -1,3 +1,5 @@
 import Container from 'components/Container'
+import Row from 'components/Row'
+import Col from 'components/Col'
 
-export default { Container }
+export default { Container, Row, Col }
